requirements/dev.txt
-r test.txt

ipython
ipdb

pip
setuptools
wheel

flake8
flake8-builtins
flake8-bugbear
flake8-mutable
flake8-comprehensions
pep8-naming
dlint
doc8

rope
isort

Sphinx

werkzeug[watchdog]
termcolor
flask-shell-ipython
